Task 1 modification

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -6,20 +6,20 @@
  * @h: A pointer to the head of the list_t list.
  *
  * Return: The number of nodes in the list.
-*/ 
+*/
 size_t print_list(const list_t *h)
 {
-size_t nodes = 0;
+unsigned int nodes = 0;
 
-while (h != NULL) 
+while (h != NULL)
 {
 if (h->str == NULL)
 {
-printf("[%zu] (nil)\n", nodes);
+printf("[%u] (nil)\n", nodes); 
 }
 else
 {
-printf("[%zu] %s\n", nodes, h->str);
+printf("[%u] %s\n", nodes, h->str);
 }
 
 nodes++;
